+++ packages/design-system/src/components/Skeleton/index.ts
@@ -1,16 +1,8 @@
+import SkeletonButton from './variations/SkeletonButton';
+import SkeletonButtonIcon from './variations/SkeletonButtonIcon';
+import SkeletonHeading from './variations/SkeletonHeading';
+import SkeletonParagraph from './variations/SkeletonParagraph';
 import Skeleton from './Skeleton';
-import SkeletonButton from './variations/Skeleton.Button';
-import SkeletonHeading from './variations/Skeleton.Heading';
-import SkeletonIconButton from './variations/Skeleton.IconButton';
 
-const SkeletonComponent = Skeleton as typeof Skeleton & {
-	Button: typeof SkeletonButton;
-	Heading: typeof SkeletonHeading;
-	Icon: typeof SkeletonIconButton;
-};
-
-SkeletonComponent.Button = SkeletonButton;
-SkeletonComponent.Heading = SkeletonHeading;
-SkeletonComponent.Icon = SkeletonIconButton;
-
-export default SkeletonComponent;
+export default Skeleton;
+export { SkeletonHeading, SkeletonButtonIcon, SkeletonButton, SkeletonParagraph };