+++ packages/design-system/src/components/Tag/Tag.tsx
@@ -1,12 +1,36 @@
-import React from 'react';
+import React, { forwardRef, Ref } from 'react';
 
-import * as S from './Tag.style';
+import { TagProps as PrimitiveTagProps } from './primitive';
+import { TagDefault, TagInformation, TagSuccess, TagWarning, TagDestructive } from './variations';
 
-export type TagProps = React.PropsWithChildren<any>;
+enum TagVariant {
+	default = 'default',
+	information = 'information',
+	success = 'success',
+	warning = 'warning',
+	destructive = 'destructive',
+}
 
-// @see https://www.selbekk.io/blog/2020/05/forwarding-refs-in-typescript/
-const Tag = React.forwardRef<HTMLSpanElement, TagProps>((props, ref) => {
-	return <S.Tag ref={ref} {...props} />;
+type TagProps = Omit<PrimitiveTagProps, 'className'> & {
+	/**
+	 * Tag variation depending on its semantic
+	 */
+	variant?: keyof typeof TagVariant;
+};
+
+const Tag = forwardRef(({ variant, ...rest }: TagProps, ref: Ref<HTMLSpanElement>) => {
+	switch (variant) {
+		case TagVariant.information:
+			return <TagInformation {...rest} ref={ref} />;
+		case TagVariant.success:
+			return <TagSuccess {...rest} ref={ref} />;
+		case TagVariant.warning:
+			return <TagWarning {...rest} ref={ref} />;
+		case TagVariant.destructive:
+			return <TagDestructive {...rest} ref={ref} />;
+		default:
+			return <TagDefault {...rest} ref={ref} />;
+	}
 });
 
 export default Tag;