+++ packages/design-system/src/index.ts
@@ -1,5 +1,18 @@
 import Accordion from './components/Accordion';
-import Button from './components/Button';
+import {
+	ButtonPrimary,
+	ButtonSecondary,
+	ButtonDestructive,
+	ButtonTertiary,
+	Button,
+} from './components/Button';
+import {
+	ButtonAsLink,
+	ButtonPrimaryAsLink,
+	ButtonSecondaryAsLink,
+	ButtonDestructiveAsLink,
+	ButtonTertiaryAsLink,
+} from './components/ButtonAsLink';
 import { ButtonIcon, ButtonIconFloating, ButtonIconToggle } from './components/ButtonIcon';
 import Card from './components/Card';
 import Combobox from './components/Combobox';
@@ -14,18 +27,37 @@ import InlineEditing from './components/InlineEditing';
 import InlineMessage from './components/InlineMessage';
 import Layout from './components/Layout';
 import Link from './components/Link';
+import { LinkAsButton } from './components/LinkAsButton';
 import Loading from './components/Loading';
 import Menu from './components/Menu';
 import Modal from './components/Modal';
 import Popover from './components/Popover';
-import Skeleton from './components/Skeleton';
-import Status from './components/Status';
+import Skeleton, {
+	SkeletonParagraph,
+	SkeletonHeading,
+	SkeletonButton,
+	SkeletonButtonIcon,
+} from './components/Skeleton';
+import {
+	StatusCanceled,
+	StatusInProgress,
+	StatusWarning,
+	StatusSuccessful,
+	StatusFailed,
+	Status,
+} from './components/Status';
 import Switch from './components/Switch';
 import Stepper from './components/Stepper';
 import Tabs from './components/Tabs';
-import Tag from './components/Tag';
+import {
+	Tag,
+	TagDefault,
+	TagInformation,
+	TagSuccess,
+	TagWarning,
+	TagDestructive,
+} from './components/Tag';
 import ThemeProvider from './components/ThemeProvider';
-import Toggle from './components/Toggle';
 import Tooltip from './components/Tooltip';
 import VisuallyHidden from './components/VisuallyHidden';
 
@@ -36,9 +68,19 @@ export * from './components/Stack';
 export {
 	Accordion,
 	Button,
+	ButtonAsLink,
+	ButtonDestructive,
+	ButtonDestructiveAsLink,
 	ButtonIcon,
 	ButtonIconFloating,
+	ButtonIconToggle as Toggle,
 	ButtonIconToggle,
+	ButtonPrimary,
+	ButtonPrimaryAsLink,
+	ButtonSecondary,
+	ButtonSecondaryAsLink,
+	ButtonTertiary,
+	ButtonTertiaryAsLink,
 	Card,
 	Combobox,
 	Divider,
@@ -50,20 +92,34 @@ export {
 	IconsProvider,
 	InlineEditing,
 	InlineMessage,
-	Link,
 	Layout,
+	Link,
+	LinkAsButton,
 	Loading,
 	Menu,
 	Modal,
 	Popover,
 	Skeleton,
+	SkeletonButton,
+	SkeletonButtonIcon,
+	SkeletonHeading,
+	SkeletonParagraph,
 	Status,
-	Switch,
+	StatusCanceled,
+	StatusFailed,
+	StatusInProgress,
+	StatusSuccessful,
+	StatusWarning,
 	Stepper,
+	Switch,
 	Tabs,
 	Tag,
+	TagDefault,
+	TagDestructive,
+	TagInformation,
+	TagSuccess,
+	TagWarning,
 	ThemeProvider,
-	Toggle,
 	Tooltip,
 	VisuallyHidden,
 	themes,