+++ packages/components/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @talend/react-components
 
+## 6.44.5
+
+### Patch Changes
+
+- Updated dependencies [7af0ef8f6]
+  - @talend/design-system@2.0.0
+
 ## 6.44.4
 
 ### Patch Changes